feat: New changes and styles
