+++ Frontend-LookStyle/app/barbershops/page.tsx
@@ -1,8 +1,10 @@
-import SearchList from "./SearchList";
+import NavBar from "@/components/NavBar";
+import SearchList from "./List";
 
 export default function BarbershopsPage() {
   return (
     <>
+      <NavBar />
       <SearchList />
     </>
   );
