+++ Frontend-LookStyle/app/barbershops/schedule/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>Schedule</div>
+  )
+}
+
+export default page
