+++ Frontend-LookStyle/next.config.js
@@ -1,11 +1,14 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {}
+const nextConfig = {};
 
-module.exports = nextConfig
+module.exports = nextConfig;
 
 module.exports = {
-    images: {
-        domains: ['st4allthings4p4ci.blob.core.windows.net', 'content.listisima.com', 'todoservy.com.co'],
-    },
-}
-    
+  images: {
+    domains: [
+      "st4allthings4p4ci.blob.core.windows.net",
+      "content.listisima.com",
+      "previews.123rf.com",
+    ],
+  },
+};
